+++ crates/dragonglass_render/src/vulkan/gui.rs
@@ -343,10 +343,10 @@ impl GuiRender {
             let clip_max_y = scale_factor * clip_rect.max.y;
 
             // Make sure clip rect can fit within an `u32`.
-            let clip_min_x = clip_min_x.clamp(0.0, viewport.width as f32);
-            let clip_min_y = clip_min_y.clamp(0.0, viewport.height as f32);
-            let clip_max_x = clip_max_x.clamp(clip_min_x, viewport.width as f32);
-            let clip_max_y = clip_max_y.clamp(clip_min_y, viewport.height as f32);
+            let clip_min_x = clip_min_x.clamp(0.0, viewport.width);
+            let clip_min_y = clip_min_y.clamp(0.0, viewport.height);
+            let clip_max_x = clip_max_x.clamp(clip_min_x, viewport.width);
+            let clip_max_y = clip_max_y.clamp(clip_min_y, viewport.height);
 
             let clip_min_x = clip_min_x.round() as u32;
             let clip_min_y = clip_min_y.round() as u32;
